исправление ошибок 3,1
