+++ data.php
@@ -1,0 +1,42 @@
+<?php
+$categories = ['Работа', 'Учеба', 'Входящие', 'Домашние дела', 'Авто'];
+
+$work = [
+    [
+        'task' => 'Собеседование в IT',
+        'date' => '01.12.2019',
+        'category' => 'Работа',
+        'finish'=>false,
+    ],
+    [
+        'task' => 'Выполнить тестовое задание',
+        'date' => '25.12.2019',
+        'category' => 'Работа',
+        'finish' => false,
+    ],
+    [
+        'task' => 'Сделать задание первого раздела',
+        'date' => '21.12.2019',
+        'category' => 'Учеба',
+        'finish' => true,
+    ],
+    [
+        'task' => 'Встреча с другом',
+        'date' => '22.12.2019',
+        'category' => 'Входящие',
+        'finish' => false,
+    ],
+    [
+        'task' => 'Купить корм для кота',
+        'date' => null,
+        'category' => 'Входящие',
+        'finish' => false,
+    ],
+    [
+        'task' => 'Заказать пиццу',
+        'date' => null,
+        'category' => 'Входящие',
+        'finish' => false,
+    ],
+];
+?>
